\name{Rmapit}
\Rdversion{1.1}
\alias{Rmapit-package}
\alias{Rmapit}


\docType{package}

\title{Rmapit - a R package to facilitate the creation of geographical maps}

\description{Rmapit is a package for R that intends to facilitate the production of geographic maps,
from simple maps to advanced maps embedding graphic charts. The package will also progressively add
functions to facilitate the generation of complete map layouts including legends. Rmapit will target
users and organizations that need to produce maps in batch way like in atlas, yearbooks, or scientific 
monitoring reports.The package intends to minimize dependencies to other packages and rely as much as
possible on the core base and graphical R functions.}


\details{
\tabular{ll}{
Package: \tab Rmapit\cr
Type: \tab Package\cr
Version
: \tab 0.1\cr
Date: \tab 2014-04-19\cr
License: \tab GPL(>=2.0)\cr
LazyLoad: \tab yes\cr
}

Rmapit is a package for R that intends to facilitate the production of geographic maps, from simple maps 
to advanced maps embedding graphic charts. 

}

\author{Emmanuel Blondel <emmanuel.blondel1@gmail.com>}

\references{}

\seealso{}

\examples{}

\keyword{spatial}
\keyword{maps}
\keyword{GIS}
\keyword{atlas}
\keyword{reporting}


